Add pull-down console style buffer.
Serves as an additional scratch area while hacking.

Also because i got too happy and did not split up my changes
this commit includes removal of chdir() per buffer. I didn't like it.

And the addition of spewing out escape sequences to update the
xterm title, which in turn is caught by comawm so it updates
the open files in the WM titles properly.